Extract resource_repository.
