+++ flask_jsonapi/__init__.py
@@ -3,7 +3,7 @@ from .api import Api
 from .filters_schema import FilterField
 from .filters_schema import FilterSchema
 from .filters_schema import ListFilterField
-from .resource_repository_views import ResourceRepository
+from .resource_repositories.repositories import ResourceRepository
 from .resource_repository_views import ResourceRepositoryDetailView
 from .resource_repository_views import ResourceRepositoryListView
 from .resource_repository_views import ResourceRepositoryViewSet
