+++ tests/test_nested_repositories.py
@@ -3,6 +3,7 @@ from unittest import mock
 
 import pytest
 
+from flask_jsonapi.resource_repositories import repositories
 from flask_jsonapi import resource_repository_views
 from flask_jsonapi.nested import nested_repository
 
@@ -117,7 +118,7 @@ class TestCreatingChildrenModels:
         child_repositorium_object.create.assert_called_with(expected_parametes, **kwargs)
 
 
-class RepositoryWithBeginTransaction(resource_repository_views.ResourceRepository):
+class RepositoryWithBeginTransaction(repositories.ResourceRepository):
     children_repositories = {}
     very_important_test_value = 0
 
@@ -128,7 +129,7 @@ class RepositoryWithBeginTransaction(resource_repository_views.ResourceRepositor
         self.very_important_test_value = 0
 
 
-class RepositoryWithOutBeginTransaction(resource_repository_views.ResourceRepository):
+class RepositoryWithOutBeginTransaction(repositories.ResourceRepository):
     children_repositories = {}
     very_important_test_value = 0
